Add Interval Literal (#190)
This PR adds SQL date intervals, i.e., supports statements such the following:
 ```SQL
 SELECT * FROM t 
    WHERE t.date BETWEEN '2021-01-01' AND DATE '2021-01-01' + 30 DAYS
```
This is achieved by a new type of literal/`Expr`, with type `kExpressionLiteralInterval`. It is returned as second operand of the binary plus (or minus, respectivley) expression, where the start date ist the first one. Only checks if duration is a positive integer and the time unit is any of second(s), minute(s), hour(s), day(s), month(s), year(s). Any further sanity checks are in response of the consuming system.

closes #184 
closes #178